Setup testing for getting counties.
