--- shopsafe-backend/src/main/java/com/google/sps/Data/County.java
@@ -1,41 +1,0 @@
-// Copyright 2019 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package com.google.sps.data;
-
-/** Class contains the name, state, and fips of a county. */
-public final class County {
-
-    private final String countyName;
-    private final String stateName;
-    private final String countyFips;
-
-    public County(String countyName, String stateName, String countyFips) {
-        this.countyName = countyName;
-        this.stateName = stateName;
-        this.countyFips = countyFips;
-    }
-
-    public String getCountyName() {
-        return countyName;
-    }
-
-    public String getStateName() {
-        return stateName;
-    }
-
-    public String getCountyFips() {
-        return countyFips;
-    }
-}
