+++ shopsafe-backend/src/main/java/com/google/sps/data/LatLng.java
